Move tests to MetafixRecordTest, fix member order & newline (#294)
