+++ metafix/src/test/java/org/metafacture/metafix/MetafixRecordTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Fabian Steeg, hbz
+ * Copyright 2021, 2023 Fabian Steeg, hbz
  *
  * Licensed under the Apache License, Version 2.0 the "License";
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.metafacture.metafix;
+package org.metafacture.metafix; // checkstyle-disable-line JavaNCSS
 
 import org.metafacture.framework.StreamReceiver;
 import org.metafacture.metamorph.api.MorphBuildException;
@@ -2716,6 +2716,138 @@ public class MetafixRecordTest {
             o -> {
                 o.get().startRecord("1");
                 o.get().literal("name", "max");
+                o.get().endRecord();
+            }
+        );
+    }
+
+    @Test
+    public void setRepeatedFieldsToEntitiesAndSetEntityMemberName() {
+        MetafixTestHelpers.assertFix(streamReceiver, Arrays.asList(
+                "nothing()"
+            ),
+            i -> {
+                i.setRepeatedFieldsToEntities(true);
+                i.setEntityMemberName("*");
+
+                i.startRecord("1");
+                i.literal("name", "max");
+                i.literal("name", "mo");
+                i.endRecord();
+            },
+            o -> {
+                o.get().startRecord("1");
+                o.get().startEntity("name");
+                o.get().literal("*", "max");
+                o.get().literal("*", "mo");
+                o.get().endEntity();
+                o.get().endRecord();
+            }
+        );
+    }
+
+    @Test
+    public void setRepeatedFieldsToEntitiesAndSetEntityMemberNameWithNumericalSubfield() {
+        MetafixTestHelpers.assertFix(streamReceiver, Arrays.asList(
+                "nothing()"
+            ),
+            i -> {
+                i.setRepeatedFieldsToEntities(true);
+                i.setEntityMemberName("*");
+
+                i.startRecord("1");
+                i.startEntity("1001 ");
+                i.literal("1", "max");
+                i.literal("1", "mo");
+                i.endEntity();
+                i.endRecord();
+            },
+            (o, f) -> {
+                o.get().startRecord("1");
+                o.get().startEntity("1001 ");
+                o.get().startEntity("1");
+                o.get().literal("*", "max");
+                o.get().literal("*", "mo");
+                f.apply(2).endEntity();
+                o.get().endRecord();
+            }
+        );
+    }
+
+    @Test
+    public void setRepeatedFieldsToEntitiesWithNumericalSubfields() {
+        MetafixTestHelpers.assertFix(streamReceiver, Arrays.asList(
+                "nothing()"
+            ),
+            i -> {
+                i.setRepeatedFieldsToEntities(true);
+
+                i.startRecord("1");
+                i.startEntity("1001 ");
+                i.literal("1", "max");
+                i.literal("1", "mo");
+                i.endEntity();
+                i.endRecord();
+            },
+            (o, f) -> {
+                o.get().startRecord("1");
+                o.get().startEntity("1001 ");
+                o.get().startEntity("1");
+                o.get().literal("1", "max");
+                o.get().literal("2", "mo");
+                f.apply(2).endEntity();
+                o.get().endRecord();
+            }
+        );
+    }
+
+    @Test
+    public void setEntityMemberNameNoArrayMarkerOrEntity() {
+        MetafixTestHelpers.assertFix(streamReceiver, Arrays.asList(
+                "nothing()"
+            ),
+            i -> {
+                i.setEntityMemberName("*"); // no arrays or entities, no effect
+
+                i.startRecord("1");
+                i.startEntity("1001 ");
+                i.literal("1", "max");
+                i.literal("1", "mo");
+                i.endEntity();
+                i.endRecord();
+            },
+            o -> {
+                o.get().startRecord("1");
+                o.get().startEntity("1001 ");
+                o.get().literal("1", "max");
+                o.get().literal("1", "mo");
+                o.get().endEntity();
+                o.get().endRecord();
+            }
+        );
+    }
+
+    @Test
+    public void setEntityMemberNameWithArrayMarker() {
+        MetafixTestHelpers.assertFix(streamReceiver, Arrays.asList(
+                "nothing()"
+            ),
+            i -> {
+                i.setEntityMemberName("*");
+
+                i.startRecord("1");
+                i.startEntity("1001 []");
+                i.literal("1", "max");
+                i.literal("1", "mo");
+                i.endEntity();
+                i.endRecord();
+            },
+            o -> {
+                o.get().startRecord("1");
+                o.get().startEntity("1001 []");
+                o.get().literal("*", "max");
+                o.get().literal("*", "mo");
+                o.get().endEntity();
                 o.get().endRecord();
             }
         );
